requirements.txt
pandas==2.2.3
plotly==5.24.1
boto3==1.35.36
streamlit==1.39.0
matplotlib==3.9.2
seaborn==0.13.2
scipy>=1.13.1
python-dotenv>=1.0.1
requests>=2.32.3
PyJWT>=2.9.0
pyarrow>=17.0.0
pytest>=8.3.3
pytest-cov>=5.0.0
